found the real cause for the bzip2 related linking problem:
Including ($QT_USE_FILE}) adds an include_directories(...) for Qt include directories at first in the include path search order of the compiler. 
When an external bzip2 development package is installed, the related bzlib.h is included instead of the local copy in 3rdparty/bzip2, which lead to different function decoration. 

The solution is to not use 'include ($QT_USE_FILE})'. Then the order of include files could be set correctly. 

Switched back to static bzip2 library usage. 

svn path=/trunk/kdesupport/kdewin-installer/; revision=1153788