added dedicated start menu support for single packsge installation.
When installing single packages, we use the name and version of the installed
package to generate the start menu root entry. Otherwise the name 'KDE' and
the version number reported by kde4-config is taken.

The installer runs kwinstartmenu with the installed package name and version,
which are stored by kwinstartmenu in <install-root>/share/config/kwinstartmenurc
and are used to generate the start menu entries.

Storing the package name and version in a installation related config file
let kwinstartmenu be able to find the correct entry on further remove or
updates either from installer, kded winstartmenu module or installers.